Add completionProvider function for SQL code completion
